% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA_mutiple_group_by_clusterProfiler.R
\name{ERA_multi_group_by_clusterProfiler}
\alias{ERA_multi_group_by_clusterProfiler}
\title{Perform Multi-group Enrichment Analysis Using clusterProfiler}
\usage{
ERA_multi_group_by_clusterProfiler(
  data_file,
  gene_col = "gene",
  group_col = "group",
  species = "human",
  save_path = ".",
  qvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  simplify_cutoff = 0.7,
  save_raw = TRUE
)
}
\arguments{
\item{data_file}{A character string specifying the path to the input data file containing "gene" and "group" columns.}

\item{species}{A character string specifying the species, either "human" or "mouse". Default is "human".}

\item{save_path}{A character string specifying the path where results will be saved. Default is current directory.}

\item{qvalueCutoff}{A numeric value specifying the cutoff for the q-value. Default is 0.05.}

\item{pAdjustMethod}{A character string specifying the method to adjust p-values, such as "BH". Default is "BH".}

\item{simplify_cutoff}{A numeric value specifying the cutoff for simplifying GO terms. Default is 0.7.}

\item{save_raw}{A logical value indicating whether to save the raw results. Default is TRUE.}
}
\value{
A list containing the processed results for GO categories ("BP", "CC", "MF") and KEGG pathways,
and a comparison of the number of terms before and after simplification.
}
\description{
This function performs Gene Ontology (GO) and KEGG enrichment analysis for multiple groups on a given gene list.
The results are processed, simplified, and saved in CSV files.
}
