% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA_one_group_by_clusterProfiler.R
\name{ERA_one_group_by_clusterProfiler}
\alias{ERA_one_group_by_clusterProfiler}
\title{Perform Enrichment Analysis Using clusterProfiler}
\usage{
ERA_one_group_by_clusterProfiler(
  gene_list,
  species = "human",
  save_path = ".",
  qvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  simplify_cutoff = 0.7,
  save_raw = TRUE
)
}
\arguments{
\item{gene_list}{A character vector of gene symbols.}

\item{species}{A character string specifying the species, either "human" or "mouse". Default is "human".}

\item{save_path}{A character string specifying the path where results will be saved. Default is current directory.}

\item{qvalueCutoff}{A numeric value specifying the cutoff for the q-value. Default is 0.05.}

\item{pAdjustMethod}{A character string specifying the method to adjust p-values, such as "BH". Default is "BH".}

\item{simplify_cutoff}{A numeric value specifying the cutoff for simplifying GO terms. Default is 0.7.}

\item{save_raw}{A logical value indicating whether to save the raw results. Default is TRUE.}
}
\value{
A list containing the processed results for GO categories ("BP", "CC", "MF") and KEGG pathways,
and a comparison of the number of terms before and after simplification.
}
\description{
This function performs Gene Ontology (GO) and KEGG enrichment analysis on a given gene list.
It supports both human and mouse species. The results are processed, simplified, and saved in CSV files.
}
\details{
This function performs enrichment analysis for biological process (BP), cellular component (CC),
and molecular function (MF) categories in GO, as well as KEGG pathway enrichment. The results
are simplified by removing redundant terms, and both raw and simplified results are saved in CSV files.
}
\examples{
\dontrun{
gene_list <- c("TP53", "BRCA1", "EGFR")
result <- ERA_one_group_by_clusterProfiler(gene_list = gene_list, species = "human")
}

}
